# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HDP App OpenApi API Documentation",
        "description": "HDP App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "hdp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Sign Up a new user",
                "description": "Returns a newly registered user data",
                "operationId": "signUp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signup",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Sign In a registered user",
                "description": "Returns a newly registered user data",
                "operationId": "signIn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Profile of a registered user",
                "description": "Profile of a registered user",
                "operationId": "userProfile",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Profile of a registered user",
                "description": "Profile of a registered user",
                "operationId": "updateUserProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Logout a registered user",
                "description": "Logout a registered user",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserLoginFormRequest": {
                "title": "Sign in Form Request Fields",
                "description": "sign in request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "info@hdp.com"
                    },
                    "password": {
                        "title": "User password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRegistrationFormRequest": {
                "title": "Sign Up Form Request Fields",
                "description": "sign up request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "first_name": {
                        "title": "User First Name",
                        "description": "First name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "last_name": {
                        "title": "User Last Name",
                        "description": "Last name of the user",
                        "type": "string",
                        "example": "Komolafe"
                    },
                    "userable_type": {
                        "title": "User Type",
                        "description": "Type of the user Customer,Facility,Professional",
                        "type": "string",
                        "example": "Customer/Facility/Professional"
                    },
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "Email of the user"
                    },
                    "telephone": {
                        "title": "User telephone",
                        "description": "Telephone of the user",
                        "type": "integer",
                        "example": "234748993333"
                    },
                    "password": {
                        "title": "User password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserUpdateFormRequest": {
                "title": "User Update Form Request Fields",
                "description": "User Update request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "first_name": {
                        "title": "User First Name",
                        "description": "First name of the user",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "middle_name": {
                        "title": "User Middle Name",
                        "description": "Middle name of the user",
                        "type": "string",
                        "example": "Busayo"
                    },
                    "last_name": {
                        "title": "User Last Name",
                        "description": "Last name of the user",
                        "type": "string",
                        "example": "Komolafe"
                    },
                    "type": {
                        "title": "User Type",
                        "description": "Type of the user Customer,Facility,Professional",
                        "type": "string",
                        "example": "Customer/Facility/Professional"
                    },
                    "gender": {
                        "title": "User Gender",
                        "description": "Gender of the user Male,Female",
                        "type": "string",
                        "example": "Male/Female"
                    },
                    "email": {
                        "title": "User email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "Email of the user"
                    },
                    "telephone": {
                        "title": "User telephone",
                        "description": "Telephone of the user",
                        "type": "integer",
                        "example": "234748993333"
                    },
                    "location_address_house_number": {
                        "title": "User Location Address",
                        "description": "Location Address of the user",
                        "type": "string",
                        "example": "No 2 Jesus Street near schneider's office"
                    },
                    "location_address_street_name": {
                        "title": "User Location Street Address",
                        "description": "Location Street Address of the user",
                        "type": "string",
                        "example": "No 2 Jesus Street near schneider's office"
                    },
                    "location_address_suburb": {
                        "title": "User Location Address Surburb",
                        "description": "Location Address Surburb of the user",
                        "type": "string",
                        "example": "No 2 Jesus Street near schneider's office"
                    },
                    "location_address_city": {
                        "title": "User Location City",
                        "description": "Location City of the user",
                        "type": "string",
                        "example": "Abuja"
                    },
                    "location_address_state": {
                        "title": "User Location State",
                        "description": "Location State of the user",
                        "type": "string",
                        "example": "Abuja"
                    },
                    "gps_location_name": {
                        "title": "User Location Name",
                        "description": "Location Name of the user",
                        "type": "string",
                        "example": "Abuja"
                    },
                    "gps_long": {
                        "title": "User GPS Long",
                        "description": "GPS Long of the user",
                        "type": "string",
                        "example": "6.444"
                    },
                    "gps_lat": {
                        "title": "User GPS Lat",
                        "description": "GPS Lat of the user",
                        "type": "string",
                        "example": "9.0333"
                    },
                    "gps_city": {
                        "title": "User GPS City",
                        "description": "GPS City of the user",
                        "type": "string",
                        "example": "Abuja"
                    },
                    "gps_state": {
                        "title": "User GPS State",
                        "description": "GPS State of the user",
                        "type": "string",
                        "example": "FCT"
                    },
                    "bank_name": {
                        "title": "User Bank Name",
                        "description": "Bank Name of the user",
                        "type": "string",
                        "example": "First Bank"
                    },
                    "bank_account_name": {
                        "title": "User Bank Account Name",
                        "description": "Bank Account Name of the user",
                        "type": "string",
                        "example": "Schneider Komolafe"
                    },
                    "bank_account_number": {
                        "title": "User Bank Account Number",
                        "description": "Bank Account Number of the user",
                        "type": "string",
                        "example": "002342222211"
                    },
                    "paystack_bank_code": {
                        "title": "User Paystack Bank code Number",
                        "description": "Paystack Bank code Number of the user",
                        "type": "string",
                        "example": "230333"
                    },
                    "paystack_bank_account_verification_details": {
                        "title": "User Paystack Bank Verification Details",
                        "description": "Paystack Bank Verification Details of the user",
                        "type": "string",
                        "example": "Details"
                    },
                    "paystack_recipient_code": {
                        "title": "User Paystack Reciepent Code",
                        "description": "Paystack Reciepent Code of the user",
                        "type": "string",
                        "example": "Details"
                    },
                    "paystack_recipient_details": {
                        "title": "User Paystack Reciepent Details",
                        "description": "Paystack Reciepent Details of the user",
                        "type": "string",
                        "example": "Details"
                    },
                    "profile_image": {
                        "title": "User Profile Image",
                        "description": "Profile Image of the user",
                        "type": "string",
                        "example": "Image"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}